Split visor api.go (4146 lines) into 10 topic files
Pure reorganization — no code changes:
- api.go (438): interface + type definitions
- api_visor.go (295): overview, health, shutdown
- api_apps.go (640): app management, VPN/skysocks control
- api_transport.go (321): transport CRUD, discovery
- api_routing.go (216): routing rules, route groups, mux
- api_ping.go (677): ping, bandwidth test, proxy test
- api_dmsg.go (701): DMSG ping, bandwidth, HTTP proxy
- api_services.go (342): service health, SD queries, config
- api_network.go (329): connect/disconnect, HTTP ports, reinit
- api_tps.go (334): embedded TPS operations, UI server
